move reducer files to directory
